requirements_osx.txt
absl-py==1.4.0
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiorwlock==1.3.0
aiosignal==1.3.1
ansi2html==1.8.0
antlr4-python3-runtime==4.9.3
anyio==3.7.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
blessed==1.20.0
bokeh==3.2.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorful==0.5.5
comm==0.1.3
contourpy==1.2.0
critdd==0.0.4.dev0
croniter==1.3.15
cycler==0.12.1
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-mantine-components==0.12.1
dash-table==5.0.0
dateutils==0.6.12
debugpy==1.6.7
decorator==5.1.1
deepdiff==6.3.0
defusedxml==0.7.1
distlib==0.3.6
dm-tree==0.1.8
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.98.0
fastjsonschema==2.17.1
filelock==3.12.0
Flask==2.2.5
flatbuffers==23.5.26
fonttools==4.46.0
fqdn==1.5.1
frozendict==2.3.8
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.3
gast==0.4.0
gmpy2==2.1.2
google-api-core==2.11.1
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
gpustat==1.1
grpcio==1.56.2
h11==0.14.0
h5py==3.9.0
huggingface-hub==0.20.3
hydra-core==1.3.2
hyperopt==0.2.7
idna==3.4
iniconfig==2.0.0
inquirer==3.1.3
ipykernel==6.23.1
ipympl==0.9.3
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.13
jaxlib==0.4.13
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==8.2.0
jupyter-console==6.6.3
jupyter_core==5.3.0
jupyter-events==0.6.3
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kaggle==1.5.13
keras==2.13.1
kiwisolver==1.4.5
liac-arff==2.5.0
libclang==16.0.6
lightning==2.0.5
lightning-cloud==0.5.37
lightning-utilities==0.8.0
linkify-it-py==2.0.2
llvmlite==0.40.1
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
minio==7.1.15
mistune==2.0.5
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
neural-tangents==0.6.2
notebook==6.5.4
notebook_shim==0.2.3
numba==0.57.1
numpy==1.24.3
nvidia-ml-py==11.525.112
oauthlib==3.2.2
omegaconf==2.3.0
opencensus==0.11.2
opencensus-context==0.1.3
openml==0.14.1
opt-einsum==3.3.0
ordered-set==4.1.0
overrides==7.3.1
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
param==1.13.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pip==23.1.2
platformdirs==3.5.1
plotly==5.15.0
pluggy==1.4.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-spy==0.3.14
py4j==0.10.9.7
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyChart.JS==0.3.0
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
PyJWT==2.7.0
pynndescent==0.5.10
pyparsing==3.1.1
pyperclip==1.8.2
pyrsistent==0.19.3
PySocks==1.7.1
pytest==8.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-json-logger==2.0.7
python-multipart==0.0.6
python-slugify==8.0.1
pytorch-lightning==2.0.5
pytz==2023.3.post1
pyviz-comms==2.3.2
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
ray==2.6.1
readchar==4.0.5
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rsa==4.9
rtdl_revisiting_models==0.0.2
safetensors==0.4.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.0
Send2Trash==1.8.2
setuptools==67.7.2
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.27.0
starsessions==1.3.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.1
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
tf2jax==0.3.4
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.15.1
tomli==2.0.1
torch==2.0.1
torch-geometric==2.3.1
torchmetrics==0.11.4
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.37.1
tune-sklearn==0.4.5
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.2
umap-learn==0.5.3
uri-template==1.2.0
urllib3==1.26.16
uvicorn==0.22.0
virtualenv==20.21.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
websockets==11.0.3
Werkzeug==2.2.3
wheel==0.40.0
widgetsnbextension==4.0.7
wrapt==1.15.0
xgboost==1.7.6
xmltodict==0.13.0
xyzservices==2023.7.0
yarl==1.9.2
